+++ test/Scoop-Decompress.Tests.ps1
@@ -3,6 +3,7 @@
 . "$psscriptroot\..\lib\unix.ps1"
 . "$psscriptroot\..\lib\install.ps1"
 . "$psscriptroot\..\lib\manifest.ps1"
+. "$psscriptroot\..\lib\versions.ps1"
 
 $isUnix = is_unix